requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.1.0
bcrypt==4.0.1
black==22.10.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
dictdiffer==0.9.0
distlib==0.3.6
ecdsa==0.18.0
fastapi==0.85.0
fastapi-pagination==0.10.0
filelock==3.8.0
flake8==5.0.4
freezegun==1.2.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
identify==2.5.7
idna==3.4
iniconfig==1.1.1
iso8601==1.1.0
isort==5.10.1
mccabe==0.7.0
mypy==0.982
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
pre-commit-hooks==4.3.0
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.10.2
pyflakes==2.5.0
PyMySQL==1.0.2
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.3
pytest-asyncio==0.20.1
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.4
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tortoise-orm==0.19.2
types-passlib==1.7.7
types-python-jose==3.3.4
types-urllib3==1.26.25.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.16.6
